fix: resolve type comparison error in txmempool.cpp
- Add static_cast to fix sign-compare error between GetSizeWithDescendants() and entry_size
- Fixes arm-linux-build CI failure
- Aligns with Bitcoin PR #23962 intent to standardize transaction size types

This resolves the CI error:
error: comparison of integer expressions of different signedness: 'int64_t' and 'uint64_t'

+++ src/txmempool.cpp
@@ -211,7 +211,7 @@ bool CTxMemPool::CalculateAncestorsAndCheckLimits(size_t entry_size,
         staged_ancestors.erase(stage);
         totalSizeWithAncestors += stageit->GetTxSize();
 
-        if (stageit->GetSizeWithDescendants() + entry_size > limitDescendantSize) {
+        if (stageit->GetSizeWithDescendants() + static_cast<uint64_t>(entry_size) > limitDescendantSize) {
             errString = strprintf("exceeds descendant size limit for tx %s [limit: %u]", stageit->GetTx().GetHash().ToString(), limitDescendantSize);
             return false;
         } else if (stageit->GetCountWithDescendants() + entry_count > limitDescendantCount) {
